Some mask fixes for 001
I realized that some transitions involving lines of width 1 powered by 001 Game Creator looked ugly not because of 001, but because I was missing some important points when making the masks.

With this update, I made some changes on the masks to make this transitions beautiful, as they are supposed to be. For this reason, 001's users can now benefit from more neat transitions.